+++ src/commands/delete.js
