+++ src/commands/delete.test.js
