+++ src/commands/deploy.js
