+++ src/commands/deploy.test.js
