+++ src/commands/list.test.js
