+++ src/commands/new.js
