+++ src/commands/new.test.js
